+++ common/cache/init.go
@@ -1,6 +1,9 @@
 package cache
 
-import "github.com/redis/go-redis/v9"
+import (
+	"context"
+	"github.com/go-redis/redis/v8"
+)
 
 var RDB = InitRedisDB()
 
@@ -9,5 +12,22 @@ func InitRedisDB() *redis.Client {
 		Addr:     "localhost:6379",
 		Password: "", // no password set
 		DB:       0,  // use default DB
+	})
+}
+
+var RdbVCid *redis.Client // video -> comment 一对多
+var RdbCVid *redis.Client // comment -> video 一对一
+var Ctx = context.Background()
+
+func InitRedis() {
+	RdbCVid = redis.NewClient(&redis.Options{
+		Addr:     "127.0.0.1",
+		Password: "xx",
+		DB:       11,
+	})
+	RdbVCid = redis.NewClient(&redis.Options{
+		Addr:     "127.0.0.1",
+		Password: "xx",
+		DB:       22,
 	})
 }
