+++ common/models/usercomment.go
@@ -1,5 +1,3 @@
-// 暂存
-
 package models
 
 import (
