--- service/rpc-user-operator/rpc-user-comment/cache/init.go
@@ -1,15 +1,0 @@
-package cache
-
-//
-//import "github.com/go-redis/redis/v8"
-//
-//var RdbVCid *redis.Client // video -> comment
-//var RdbCVid *redis.Client // comment -> video
-//
-//func InitRedis() {
-//	RdbCVid = redis.NewClient(&redis.Options{
-//		Addr:     "127.0.0.1",
-//		Password: "xx",
-//		DB:       1,
-//	})
-//}
